key serialization/deserialization
