+++ src/crypto/mod.rs
@@ -1,0 +1,3 @@
+/// Abstracted cryptographic primitives.
+/// It is not recommended to use these primitives directly, instead use a pre-defined cryptosystem.
+pub mod primitives;
